+++ src/xrGame/Inventory.h
@@ -90,7 +90,7 @@ public:
 	// Ищет на поясе или в рюкзаке IItem с указанным именем (cName())
 	PIItem Get(LPCSTR name, bool bSearchRuck) const;
 	// Ищет на поясе или в рюкзаке IItem с указанным именем (id)
-	PIItem Get(const u16 id, bool bSearchRuck) const;
+	PIItem Get(const u32 id, bool bSearchRuck) const;
 	// Ищет на поясе или в рюкзаке IItem с указанным CLS_ID
 	PIItem Get(CLASS_ID cls_id, bool bSearchRuck) const;
 	PIItem GetAny(LPCSTR name) const; //search both (ruck and belt)
